Skip building manylinux1 wheel for Python 3.4

+++ build_many_linux.sh
@@ -3,8 +3,6 @@ set -e
 
 # Inspired by https://github.com/pypa/python-manylinux-demo/blob/a615d78e5042c01a03e1bbb1ca78603c90dbce1f/travis/build-wheels.sh
 
-# To build 32bit wheels, run:
-# docker run --rm -v $PWD:/data -e "UAMQP_REBUILD_PYX=True" local/manylinux_crypto32 /data/build_many_linux.sh
 # To build 64bit wheels, run:
 # docker run --rm -v $PWD:/data -e "UAMQP_REBUILD_PYX=True" local/manylinux_crypto64 /data/build_many_linux.sh
 
@@ -15,15 +13,16 @@ export OPENSSL_ROOT_DIR="/opt/pyca/cryptography/openssl"
 # Build the wheels
 pushd /data;
 for PYBIN in /opt/python/*/bin; do
-	$PYBIN/pip install cython==0.28.5 wheel;
-	$PYBIN/python setup.py bdist_wheel -d /wheelhouse;
-	rm -rf build/
+        if [[ "$PYBIN" =~ "cp34" ]]; then continue; fi; # Skip Python 3.4
+        $PYBIN/pip install cython==0.28.5 wheel;
+        $PYBIN/python setup.py bdist_wheel -d /wheelhouse;
+        rm -rf build/
 done;
 popd;
 
 # Repair the wheels
 for WHL in /wheelhouse/*; do
-	auditwheel repair $WHL -w /data/wheelhouse/;
+        auditwheel repair $WHL -w /data/wheelhouse/;
 done;
 
 # Set up env vars to run live tests - otherwise they will be skipped.
@@ -40,6 +39,7 @@ export IOTHUB_SAS_KEY=""
 
 # Test the wheels
 for PYBIN in /opt/python/*/bin; do
+        if [[ "$PYBIN" =~ "cp34" ]]; then continue; fi; # Skip Python 3.4
         $PYBIN/pip install "certifi>=2017.4.17" "six~=1.0" "enum34>=1.0.4" "pytest" "pylint";
         $PYBIN/pip install uamqp --no-index -f /data/wheelhouse;
         $PYBIN/python -c 'import uamqp;print("*****Importing uamqp from wheel successful*****")';
